feat/LJJ(#46): query component분리 및 리뷰 반영
