+++ src/components/Main/Profile/UnConnected.tsx
@@ -24,7 +24,9 @@ export default function UnConnectedProfile() {
             </div>
           </S.EachImageContainer>
           <S.CoupleShortIntroduction>
-            <h4>{myInfo?.nickname}</h4>
+            <h4>
+              {myInfo ? myInfo.nickname : <>설정에서 이름을 설정해 주세요.</>}
+            </h4>
             <p>
               소중한 추억을 함께 기록하기 위해 파트너와 연결해주세요.
               <br /> 연결 후에는 다이어리 작성, 사진 공유 등 더 많은 기능을
